services/user-org/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0.post1
structlog==23.2.0
aiokafka==0.10.0
